fix: directories and code changes
