+++ apps/frontend/app/app/layout.tsx
@@ -1,22 +1,27 @@
 import { AppSidebar } from "@/components/app-sidebar";
 import { SidebarProvider } from "@/components/ui/sidebar";
-
+import "../globals.css";
 export default function AppLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
-    <SidebarProvider defaultOpen={false}>
+    <html lang="en">
+      <body>
+      <SidebarProvider defaultOpen={false}>
       <div className="relative flex min-h-screen w-full">
         <AppSidebar />
         <div
           className="relative flex-1 w-full pt-16"
-          style={{ marginLeft: "60px" }}
         >
           {children}
         </div>
       </div>
     </SidebarProvider>
+      </body>
+
+    </html>
+
   );
 }
