+++ apps/frontend/components/Join_the_community/JoinTheCommunity.tsx
