+++ apps/frontend/components/Join_the_community/TestimonialComponent.tsx
