+++ apps/frontend/components/hero_section_component/HeroSection.tsx
@@ -4,7 +4,6 @@ export default function HeroSection() {
   return (
     <section className="bg-primary-foreground text-white py-16 px-4">
       <div className="container mx-auto flex flex-col md:flex-row items-center justify-between gap-8">
-        {/* Text Column */}
         <div className="md:w-1/2 flex flex-col items-center md:items-start text-center md:text-left">
           <h1 className="font-inter font-bold text-[32px] leading-[38.73px] tracking-[0] mb-4">
             Conquer the Dojo
@@ -14,14 +13,12 @@ export default function HeroSection() {
             Starknet.
           </p>
           <div className="flex flex-col md:flex-row gap-4">
-            {/* Primary Button */}
             <a
-              href="/tatami-gui"
+              href="/app"
               className="bg-primary text-primary-foreground font-semibold py-3 px-6 rounded-md transition-colors hover:bg-yellow-500"
             >
               Try Tatami
             </a>
-            {/* Secondary Button */}
             <a
               href="/learn-more"
               className="border border-primary text-primary font-semibold py-3 px-6 rounded-md transition-colors hover:bg-primary-foreground"
@@ -30,8 +27,6 @@ export default function HeroSection() {
             </a>
           </div>
         </div>
-
-        {/* Image Column */}
         <div className="md:w-1/2">
           <img
             src="/hero-illustration.png"
